Move service_schema functions to its own packages
