+++ mtpylon/service_schema/__init__.py
@@ -28,16 +28,18 @@ from .constructors import (
     DestroyAuthKeyRes,
     DestroySessionRes
 )
-from .req_pq import req_pq
-from .req_pq_multi import req_pq_multi
-from .req_DH_params import req_DH_params
-from .set_client_DH_params import set_client_DH_params
-from .rpc_drop_answer import rpc_drop_answer
-from .get_future_salts import get_future_salts
-from .ping import ping
-from .ping_delay_disconnect import ping_delay_disconnect
-from .destroy_auth_key import destroy_auth_key
-from .destroy_session import destroy_session
+from .functions import (
+    req_pq,
+    req_pq_multi,
+    req_DH_params,
+    set_client_DH_params,
+    rpc_drop_answer,
+    get_future_salts,
+    ping,
+    ping_delay_disconnect,
+    destroy_auth_key,
+    destroy_session
+)
 
 service_schema = Schema(
     constructors=[
