+++ mtpylon/service_schema/functions/destroy_auth_key.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from .constructors import DestroyAuthKeyOk, DestroyAuthKeyRes
+from ..constructors import DestroyAuthKeyOk, DestroyAuthKeyRes
 
 
 async def destroy_auth_key() -> DestroyAuthKeyRes:
