+++ mtpylon/service_schema/functions/destroy_session.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-from ..utils import long
-from .constructors import DestroySessionRes, DestroySessionOk
+from ...utils import long
+from ..constructors import DestroySessionRes, DestroySessionOk
 
 
 async def destroy_session(session_id: long) -> DestroySessionRes:
