+++ mtpylon/service_schema/functions/ping_delay_disconnect.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-from ..utils import long
-from .constructors import Pong
+from ...utils import long
+from ..constructors import Pong
 
 
 async def ping_delay_disconnect(ping_id: long, disconnect_delay: int) -> Pong:
