+++ mtpylon/service_schema/functions/rpc_drop_answer.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-from ..utils import long
-from .constructors import RpcDropAnswer, RpcAnswerUnknown
+from ...utils import long
+from ..constructors import RpcDropAnswer, RpcAnswerUnknown
 
 
 async def rpc_drop_answer(req_msg_id: long) -> RpcDropAnswer:
